+++ CHANGE_LOG.md
@@ -1,3 +1,13 @@
+# version 0.11.1-preview.1
+## Refactored OpenId public key fetching to allow for caching
+**Type of change:** engineering    
+**Customer impact:** low
+
+# version 0.11.1-preview.0
+## Rules Model updates to support extensiblity features
+**Type of change:** new feature    
+**Customer impact:** low
+
 # version 0.10.1
 ## Support for json-ld proofs
 **Type of change:** new feature